LayoutTests/discovery/discovery-zeroconf-unavailable-expected.txt

Add request
WebView network services request permission 2 service(s) (origin "file://")
	Services:
	 - 2 allowed
	 - 0 denied
	 - 2 online
	 - 0 offline
Tests that unavailable event is send when a service requested is removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
* register default error callback
* sent network services request
* got reply: 2 service(s) on 2 services available
* services[0]: zeroconf:_test-unavailable._tcp
* services[1]: zeroconf:_test-unavailable._tcp
* register onunavailable for service 0
* register onunavailable for service 1
* service unavailable : TestRunner1
